final add_item button (without update database)

+++ HungryJava/app/src/main/java/com/example/hungryjava/Item.java
@@ -2,20 +2,20 @@ package com.example.hungryjava;
 
 public class Item {
     private String name;
-    private Double quantity;
+    private int quantity;
 
-    public Item(String name, Double quantity)
+    public Item(String name, int quantity)
     {
         this.name = name;
         this.quantity = quantity;
     }
 
-    public void setQuantity(Double newAmount)
+    public void setQuantity(int newAmount)
     {
         this.quantity=newAmount;
     }
 
-    public double getQuantity()
+    public int getQuantity()
     {
         return quantity;
     }
